Added runtime option to skip cairo and use svgwrite instead
Added a boolean option --svgwrite which uses svgwrite library
instead of cairo for saving svg files. When -svgwrite is enabled,
the program has no rumtime dependencies on cairo.

The svgwrite version seems to be functionally equivalent.
Slower because converting floats to strings is now in python.
But lacking some visual artifacts on test image.